fix: validate api keys before csrf bypass
