+++ backend/src/middleware/csrfConfig.ts
@@ -1,5 +1,6 @@
 import { Request } from "express";
 import crypto from "crypto";
+import { isApiKeyAuthorized } from "../utils/apiKeyAuth";
 
 export const CSRF_SECRET =
   process.env.CSRF_SECRET || crypto.randomBytes(32).toString("hex");
@@ -34,12 +35,5 @@ export const isRssManagementRequest = (req: Request): boolean => {
   );
 };
 
-export const isApiKeyRequest = (req: Request): boolean => {
-  return Boolean(
-    req.headers["x-api-key"] ||
-      req.headers.authorization?.startsWith("ApiKey ")
-  );
-};
-
 export const shouldSkipCsrfProtection = (req: Request): boolean =>
-  !isRssManagementRequest(req) && isApiKeyRequest(req);
+  !isRssManagementRequest(req) && isApiKeyAuthorized(req);
